relax thresholds
make it slightly less dependent on the 4.8v programmer voltage -- should now work at 5.0v as well.

the digitalRead checks failed because it's all designed for a 5v LED drive voltage but for the attiny to run at 3.3v... for some reason. thanks, me. but yeah we can't check that here.